\name{I.0}
\title{Zeroth Order Bessel Function of the First Kind}
\alias{I.0}

\description{
Returns an approximation of the zeroth order Bessel function of the first kind, evaluated at a specified real number. 
}

\usage{
I.0(x)
}

\arguments{
\item{x}{non-negative numerical value at which to evaluate the Bessel function.}
}

\value{
Returns an approximation of the zeroth order Bessel function of the first kind evaluated at a specified real number.
}

\details{
The value returned is a polynomial approximation based on the algorithm of Abramowitz and Stegun (1970).  For a reference, see below.
}

\references{
Abramowitz, M. and Stegun, I. (1970). Handbook of Mathematical Functions.  National Bureau of Standards, New York.
}

\seealso{
\link{I.1}, \link{I.p}, \link{A1}.
}

\keyword{math}

